move MediaType to constants
