+++ src/arr.ts
@@ -2,7 +2,7 @@ import chalk from "chalk";
 import ms from "ms";
 import { join as posixJoin } from "node:path/posix";
 import { URLSearchParams } from "node:url";
-import { SCENE_TITLE_REGEX } from "./constants.js";
+import { MediaType, SCENE_TITLE_REGEX } from "./constants.js";
 import { CrossSeedError } from "./errors.js";
 import { Caps } from "./indexers.js";
 import { Label, logger } from "./logger.js";
@@ -13,7 +13,6 @@ import {
 	cleanseSeparators,
 	getApikey,
 	isTruthy,
-	MediaType,
 	sanitizeUrl,
 	stripMetaFromName,
 } from "./utils.js";
